requirements.txt
flask
web3
requests
python-dotenv
streamlit
pandas
chardet
openzeppelin-contracts  # For compilation